Fix uibei to return 0 for appropriate conditions
The exception was being raised because the method was proceeding with
the calculation even though the chemical potential was greater than
the lower limit of integration. In this case, the result should
explicitly be zero. I expanded the docstring of the method to note the
integral form of the upper incomplete Bose-Einstein integral as well as
the condition for which the result should be zero.

Closes issue #2.