Reorganize documentation
I streamlined the documentation in the `README.md` file to be more of
a quickstart; the contents of this file will be displayed when people
visit the github repository page and I figured these readers will be
able to get along with the bare minimum of documentation. I will use
github pages to host the full documentation and so I moved the
detailed documentation into the `doc` directory. The contents of
`index.rst` were copied directly from `README.md` and so the result is
not properly formatted.