Properly specify dependencies
I was trying to indicate that the `physicalproperty` package was located
in my (jrsmith3) binstar channel. However, the syntax I used was
incorrect. In any event, the new meta.yaml file results in the `conda
build` command working. I think it works because the jrsmith3 channel is
specified in my `.condarc` file. I think that if the jrsmith3 channel
was not specified in `.condarc`, I would have to execute the following
command to get this package to build properly:

```
conda build -c jrsmith3 conda.recipe
```

It turns out that even with the jrsmith3 channel missing from
`.condarc`, the build command still works.

Closes #27.