Import only the class under test
I have a lot of tests that are similar across the three classes in
this module. This approach prevents mistakes when I am copying and
pasting from one test suite to another. Specifically, I might
accidentally forget to update the call to the constructor in one
of the tests and the tests just pass without alerting me of the
mistake.